Add AddAdapterButton with tooltip support to management page header
Moves the management page title and subtitle into a flex layout to accommodate a new AddAdapterButton component in the header. The button is positioned on the right side using flexbox and will be used to trigger adapter creation. Tooltip components are imported to support hover feedback for the button, enhancing user experience.